Replaced functions in GameManager
